Combine duplicated family detail screen tests
